pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift"
version = "0.1.0"
description = "Organize your Mac's files with AI"
authors = [{name = "You", email = "you@example.com"}]
requires-python = ">=3.12"
dependencies = [
    "typer[all]==0.9.0",
    "click==8.1.7",
    "rich>=13.7.0",
    "textual>=0.57.1",
    "numpy<2.0.0",
    "httpx<0.28.0",
    "chromadb==0.5.0",
    "openai==1.14.0",
    "pydantic>=2.7.0",
    "pdfplumber==0.10.3",
    "python-docx==0.8.11",
    "pillow>=10.3.0",
    "python-dotenv==1.0.0",
    "pyyaml>=6.0.1",
    "sentence-transformers>=2.2.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "open_clip_torch>=0.17.0",
    "GitPython>=3.1.30",
    "hdbscan>=0.8.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
drift = "drift.__main__:main"
